Rename package to outboundmessage
