+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/outboundmessage/OutBoundMessage.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.messagebatch;
+package org.hisp.dhis.outboundmessage;
 
 /*
  * Copyright (c) 2004-2016, University of Oslo
@@ -28,10 +28,59 @@ package org.hisp.dhis.messagebatch;
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import java.util.Set;
+
 /**
- * @author Halvdan Hoem Grelland
- */
-public interface OutboundMessageBatchService
+* @author Zubair <rajazubair.asghar@gmail.com>
+*/
+
+public class OutBoundMessage
 {
-    MessageResponseSummary sendBatch( OutboundMessageBatch outboundBatch );
+    private String text;
+    
+    private Set<String> recipients;
+    
+    private String subject;
+
+    public OutBoundMessage()
+    {
+    }
+   
+    public OutBoundMessage( String text, Set<String> recipients, String subject )
+    {
+        super();
+        this.text = text;
+        this.recipients = recipients;
+        this.subject = subject;
+    }
+
+    public String getText()
+    {
+        return text;
+    }
+
+    public void setText( String text )
+    {
+        this.text = text;
+    }
+
+    public Set<String> getRecipients()
+    {
+        return recipients;
+    }
+
+    public void setRecipients( Set<String> recipients )
+    {
+        this.recipients = recipients;
+    }
+
+    public String getSubject()
+    {
+        return subject;
+    }
+
+    public void setSubject( String subject )
+    {
+        this.subject = subject;
+    }
 }
