+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/message/DefaultProgramMessageService.java
@@ -35,7 +35,6 @@ import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.common.IllegalQueryException;
 import org.hisp.dhis.common.DeliveryChannel;
 import org.hisp.dhis.message.MessageSender;
-import org.hisp.dhis.messagebatch.OutboundMessageBatchService;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramInstance;
@@ -43,11 +42,11 @@ import org.hisp.dhis.program.ProgramInstanceService;
 import org.hisp.dhis.program.ProgramService;
 import org.hisp.dhis.program.ProgramStageInstance;
 import org.hisp.dhis.program.ProgramStageInstanceService;
-import org.hisp.dhis.messagebatch.BatchResponseStatus;
-import org.hisp.dhis.messagebatch.MessageBatchCreatorService;
-import org.hisp.dhis.messagebatch.MessageBatchStatus;
-import org.hisp.dhis.messagebatch.MessageResponseSummary;
-import org.hisp.dhis.messagebatch.OutboundMessageBatch;
+import org.hisp.dhis.outboundmessage.BatchResponseStatus;
+import org.hisp.dhis.outboundmessage.MessageBatchCreatorService;
+import org.hisp.dhis.outboundmessage.MessageBatchStatus;
+import org.hisp.dhis.outboundmessage.MessageResponseSummary;
+import org.hisp.dhis.outboundmessage.OutboundMessageBatch;
 import org.hisp.dhis.trackedentity.TrackedEntityInstanceService;
 import org.hisp.dhis.user.CurrentUserService;
 import org.hisp.dhis.user.User;
