+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/pushanalysis/DefaultPushAnalysisService.java
@@ -56,7 +56,7 @@ import org.hisp.dhis.scheduling.TaskCategory;
 import org.hisp.dhis.scheduling.TaskId;
 import org.hisp.dhis.setting.SettingKey;
 import org.hisp.dhis.setting.SystemSettingManager;
-import org.hisp.dhis.messagebatch.MessageResponseStatus;
+import org.hisp.dhis.outboundmessage.MessageResponseStatus;
 import org.hisp.dhis.system.grid.GridUtils;
 import org.hisp.dhis.system.notification.NotificationLevel;
 import org.hisp.dhis.system.notification.Notifier;
